requirements.txt
Package                 Version
----------------------- -----------
annotated-types         0.7.0
anyio                   4.4.0
certifi                 2024.7.4
charset-normalizer      3.3.2
click                   8.1.7
coloredlogs             15.0.1
contourpy               1.2.1
cycler                  0.12.1
dlib                    19.24.4
dnspython               2.6.1
email_validator         2.2.0
face-recognition        1.3.0
face_recognition_models 0.3.0
fastapi                 0.111.0
fastapi-cli             0.0.4
flatbuffers             24.3.25
fonttools               4.53.0
h11                     0.14.0
httpcore                1.0.5
httptools               0.6.1
httpx                   0.27.0
humanfriendly           10.0
idna                    3.7
Jinja2                  3.1.4
kiwisolver              1.4.5
markdown-it-py          3.0.0
MarkupSafe              2.1.5
matplotlib              3.9.1
mdurl                   0.1.2
mpmath                  1.3.0
mysql-connector-python  9.0.0
nudenet                 3.4.2
numpy                   1.26.4
onnxruntime             1.18.1
opencv-python-headless  4.10.0.84
orjson                  3.10.6
packaging               24.1
pillow                  10.4.0
pip                     24.1.1
protobuf                5.27.2
pydantic                2.8.2
pydantic_core           2.20.1
Pygments                2.18.0
pyparsing               3.1.2
python-dateutil         2.9.0.post0
python-dotenv           1.0.1
python-multipart        0.0.9
PyYAML                  6.0.1
requests                2.32.3
rich                    13.7.1
setuptools              70.2.0
shellingham             1.5.4
six                     1.16.0
sniffio                 1.3.1
starlette               0.37.2
sympy                   1.12.1
typer                   0.12.3
typing_extensions       4.12.2
ujson                   5.10.0
urllib3                 2.2.2
uvicorn                 0.30.1
uvloop                  0.19.0
watchfiles              0.22.0
websockets              12.0
